+++ packages/fe-fpm-writer/src/app/index.ts
@@ -21,6 +21,8 @@ export interface FPMConfig {
      * (Note: if set to false, nothing will be done i.e. FCL is not disabled)
      */
     fcl?: boolean;
+
+    typescript?: boolean;
 }
 
 export const MIN_VERSION = '1.94.0';
@@ -91,8 +93,9 @@ export function enableFPM(basePath: string, config: FPMConfig = {}, fs?: Editor)
 
     // replace Component.js
     if (config.replaceAppComponent) {
-        const componentTemplate = join(__dirname, '../../templates/app/Component.js');
-        fs.copyTpl(componentTemplate, join(basePath, 'webapp/Component.js'), manifest['sap.app']);
+        const ext = config.typescript ? 'ts' : 'js';
+        const componentTemplate = join(__dirname, `../../templates/app/Component.${ext}`);
+        fs.copyTpl(componentTemplate, join(basePath, `webapp/Component.${ext}`), manifest['sap.app']);
     }
 
     return fs;