+++ packages/fe-fpm-writer/templates/app/Component.ts
@@ -1,0 +1,20 @@
+import AppComponent from 'sap/fe/core/AppComponent';
+
+/**
+ * @namespace <%- app.id %>
+ */
+export default class Component extends AppComponent {
+
+	public static metadata = {
+		manifest: "json"
+	};
+
+    /**
+     * The component is initialized by UI5 automatically during the startup of the app and calls the init method once.
+     * @public
+     * @override
+     */
+	//public init() : void {
+    //    super.init();
+	//}
+}