+++ packages/fe-fpm-writer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/fe-fpm-writer
 
+## 0.14.0
+
+### Minor Changes
+
+-   d351f81: Typescript support added
+
+## 0.13.6
+
+### Patch Changes
+
+-   9e6d402: Add id to page in generated custom page
+
 ## 0.13.5
 
 ### Patch Changes